requirements.txt
absl-py==2.1.0
accelerate==0.30.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.29.0.1
colorama==0.4.6
comm==0.2.2
confection==0.1.4
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
expecttest==0.1.3
fastjsonschema==2.19.1
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.62.1
huggingface-hub==0.23.0
hypothesis==5.35.1
idna==3.6
igraph==0.11.4
importlib-metadata==7.1.0
importlib-resources==6.4.0
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
jupyterlab==2.3.2
jupyterlab-pygments==0.3.0
jupyterlab-server==1.2.0
jupytext==1.16.1
kiwisolver==1.4.5
langcodes==3.3.0
lark==1.1.9
lazy-loader==0.4
librosa==0.10.1
lightning-utilities==0.11.2
llvmlite==0.41.1
looseversion==1.3.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==5.1.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.6.3
ninja==1.11.1.1
nltk==3.8.1
notebook==6.4.10
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda120==1.36.0
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvimgcodec-cu12==0.2.0.7
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
nvidia-pyindex==1.0.9
nvitop==1.3.2
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
portalocker==2.8.2
preshed==3.0.9
prettytable==3.10.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1-modules==0.4.0
pybind11==2.12.0
pybind11-global==2.12.0
pycparser==2.22
pydantic==2.6.4
pydantic-core==2.16.3
pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
pytest-flakefinder==1.1.0
pytest-rerunfailures==14.0
pytest-shard==0.1.2
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
python-hostlist==1.23.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
rsa==4.9
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==2.1.1
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx-glpi-theme==0.6
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
terminado==0.18.1
texttable==1.7.0
thinc==8.2.3
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.9.0
transformers==4.40.2
triton==2.3.1
typer==0.9.4
types-dataclasses==0.6.6
typing-extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
werkzeug==3.0.2
xdoctest==1.0.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
